+++ js/main.js
@@ -4,8 +4,5 @@
     Настройка скриптов
     ...
 */
-import './util.js';
-import './get-cards.js';
-import './get-descriptions.js';
-import './get-likes.js';
-import './get-comments.js';
+import './utils.js';
+import './data.js';
